rename types.js to propTypes.js
